+++ example/lib/interactive/pages/home/transition_type.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+enum TransitionType {
+  fadeForwards('Fade Forwards', FadeForwardsPageTransitionsBuilder()),
+  cupertino('Cupertino', CupertinoPageTransitionsBuilder()),
+  zoom('Zoom', ZoomPageTransitionsBuilder()),
+  fade('Fade Upwards', FadeUpwardsPageTransitionsBuilder()),
+  slideUp('Slide Up', OpenUpwardsPageTransitionsBuilder());
+
+  final String name;
+  final PageTransitionsBuilder builder;
+
+  const TransitionType(this.name, this.builder);
+}
+