+++ CHANGELOG.md
@@ -1,6 +1,11 @@
+## 2.0.0
+
+* Breaking Change: swipeVelocityThreshold is now double instead of GestureMeasurement (since setting swipe speed as a percentage of screen width doesn't make sense)
+* Updated Web Demo, now it provides an interactive site with instant customization of transition parameters
+
 ## 1.1.0
 
-* Fix analyzer problems
+* Fix compatibility with older versions of flutter
 
 ## 1.0.0